setup: add new ssh_host_key_types_to_generate helper to plugin API
To guard the ct_file_set_contents call so that it's really only done
if the plugin wants ssh keys.

we'll use that in a future patch to have a more abstract interface
for plugins.

Additional benefit, we can drop (or at least skip) old key types (DSA
anyone) per container os-release type/version with this.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>