format disk: set FS root uid/gid for passed through /dev/ volumes
When calling mkfs one must pass the root uid/gid parameter along side
as they are used unconditionally, but this wasn't done for the edge
case where a block device from the host was used as volume for the CT,
causing an undef warning.

Note that this code branch is not reachable currently, but that might
change. For now add a FIXME comment to mark this for removal, as we
probably do not want to format devices from /dev/ in any way (and no
user reported that this was broken, so use case seems to be
non-existent).

Fixes: d216e89 ("unprivileged: remove bad chown -R call")
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>