api: fix regression with locking start-after-create
The API now correctly forks first a worker and then locks the config
in for the actual creation, but this broke the start-after-create
feature, as that also locks the config.

Use the same approach like in qemu-server and just do the start after
the create, outside of the lock.

While this has a small race window where another API call could lock
the newly created CT again, we never really guaranteed that the start
after create param is guaranteed atomic.

Even if we want to guarantee that someday, we can still do so, but
this is a good stop-gap until then and worked fine for VMs since its
introduction.

Fixes: 7a73568 ("api: status: move config locking from API handler into worker")
Reported-by: Stefan Hanreich <s.hanreich@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Tested-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Tested-by: Fiona Ebner <f.ebner@proxmox.com>
Tested-by: Stefan Hanreich <s.hanreich@proxmox.com>