api: autocomplete rootdir storages for create, move-volume and clone
Adds a helper subroutine for enumerating storages that are enabled and
have the content type `rootdir` set, therefore supporting container
directories.

The autocompletion is added to the clone command and changed for the
create and move-volume commands, which previously suggested any storage
device. This is misleading as these commands will fail for any storage
that is not configured to store container directories.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>