network config: avoid committing invalid config on hotplug failure
If the hotplug of an interface on a lxc container fails for whatever
reason, the configuration will be broken and needs to manually fixed.

For example when adding a network interface with a bridge to a evpn vnet
(which doesn't support vlans) and we add a vlan tag, the interface will
be created even though we get an error. This will result in a broken
config (a interface without a bridge), which will cause the container to
not start anymore. Furthermore the veth interface will remain in a
`nomaster` state, which means the interface isn't connected to anything.
To solve this you would need to remove the interface manually from the
config.

To fix this we remove the logic that writes the intermediary config as
the config is wrong. This obviously reduces the consistency of the
config in some way, although that shouldn't be a problem (as it's
illegal anyway). We also need to revert to the old config in case the
new config can't be applied.

We also abort the api handler if we get an error updating the pending
config – this is not really necessary in this case, as we refrain from
writing the bad config completely. But it is nevertheless a good
practice because we won't write any other potentially bad config which
was produced during an erroneous pending config update.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>