api: return all addresses of an interface
Return all ip-addresses of an interface, not only the first one. Change
return schema to resemble the 'agent/network-get-interfaces' qemu call
reponse. This helps us making the AgentIPView more generic and display
the ip on both containers and vms.
Preserve old attributes so that we remain backwards-compatible. These
should be removed in the next version.

Fixes: #5339
Cc: Johannes Draaijer <jcdra1@gmail.com>
Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Tested-by: Daniel Herzig <d.herzig@proxmox.com>
Link: https://lore.proxmox.com/20250108143812.241488-4-g.goller@proxmox.com