backup: code cleanup: remove unused variable
Before commit 2e57a9f ("vzdump: replace diskinfo with disks array")
the variable was used and while the commit did change the assignment
of the variable, the variable is actually unused since then.

Note that '$default_mount_point' is the mountpoint of the first disk
in the '$disks' array, so the separate variable is not needed.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>