api: config: permission checks: fix check for 'nesting' feature
As reported in the community forum [0], remote migration for a container
with 'features: nesting=0' would fail with a confusing error message

> changing feature flags (except nesting) is only allowed for root@pam

This is because of a faulty check during container creation that would
only detect 'nesting=1', but not 'nesting=0'. Fix it by properly
checking if only the 'nesting' property is affected via an 'exists'
check for the hash key.

The check for deleting the feature is also broken, so fix it too.

The check for modifying the feature already works correctly.

[0]: https://forum.proxmox.com/threads/161432/

Fixes: 88ef170 ("allow nesting to be changed for VM.Allocate on unprivileged containers")
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>